Se Actualiza base de datos con respecto al modelo definido

+++ FertiWeb/src/main/java/dominio/Rol.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class Rol {
 	
-	private int codigo;
+	private Long codigo;
 	private String nombre;
 	private boolean estado;
 }
