+++ FertiWeb/src/main/java/persistencia/repositorio/FuenteRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import persistencia.entidad.FuenteEntidad;
 
-public interface FuenteRepository extends JpaRepository<FuenteEntidad, Integer> {
+public interface FuenteRepository extends JpaRepository<FuenteEntidad, Long> {
 
 }
