+++ FertiWeb/src/main/java/persistencia/repositorio/RolRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import persistencia.entidad.RolEntidad;
 
-public interface RolRepository extends JpaRepository<RolEntidad, Integer> {
+public interface RolRepository extends JpaRepository<RolEntidad, Long> {
 
 }
