+++ FertiWeb/src/main/java/persistencia/repositorio/TipoFuenteRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import persistencia.entidad.TipoFuenteEntidad;
 
-public interface TipoFuenteRepository extends JpaRepository<TipoFuenteEntidad, Integer> {
+public interface TipoFuenteRepository extends JpaRepository<TipoFuenteEntidad, Long> {
 
 }
