+++ FertiWeb/src/main/java/persistencia/repositorio/UsuarioRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import persistencia.entidad.UsuarioEntidad;
 
-public interface UsuarioRepository extends JpaRepository<UsuarioEntidad, Integer> {
+public interface UsuarioRepository extends JpaRepository<UsuarioEntidad, Long> {
 
 }
